Fixed bug #1105931.
Initialise the sw->window to zero in sheet_create. This means that
starting the join editor on the default (ie "next") match for Find
Internal Joins on the demo/gap/DEMO.0 database does not crash on
Windows.

I do not know why the lack of initialisation was so data specific and
OS specific. Valgrind reports no problems on linux and in theory this
is correct (as sw->window shouldn't be accessed until after it's been
initialised for real, due to Tk_IsMapped calls elsewhere).


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@603 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1