The defaults for -N and -n are now using the same naming conventions used in Gerald during the fastq generation steps. To do this is looks at the run folder root directory name to get the run date, machine name and run number. (These are available for use as %d, %m and %r in the format strings.)
Calibrated confidence values are now automatically included if the -qf
or -qr parameters are used (specifying the fastq filename). Note this
only works currently if the number of bases after calibration is the
same as the number before.
The calibrated confidence values are written in a CNF1 ztr chunk (in
addition to the existing CNF4 chunk for uncalibrated values) and are
rescaled to adhere to the phred scale (-10 * log10(1-P)).

Added meta-data to the confidence chunks (CNF1 and CNF4) with a SCALE
key. The value is either LO (log-odds) or PH (phred). This increases
file size somewhat as it's written once per trace, but the long-term
goal is to upgrade ZTR to support the ability to specific default
meta-data keys/values.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1202 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1