No longer iterate through tiles printing up . or ! depending on whether we encounter an error. Now it just aborts at the point of failure.
Also made the parsing code more robust as in a couple specific cases
it only wrote to stderr without actually generating a non-zero exit
code.

These mean the tool is more amenable to running in a production
pipeline. If it gets any error at all it'll be more obvious and forces
attention.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1255 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1