Overhauled the SRF indexing code.
Much of the indexing code in srf_index_hash.c has been moved over to
srf.c so it can be used by other programs.  An API has been created
too so it is now far easier to create, add to and save an index.

Added support for writing indexes in illumina2srf. Note that now if no
index is written we also write out 8 bytes of zero, indicating the
length of the index is zero. (This is required by more recent versions
of the SRF specification.)

Still to do: tools such as srf_filter should be updating the index (or
at least removing the old ones). This will now be easier to do with
these code updates.

Updated the tests to check the new illumina2srf -i option too.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1305 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1