Added starting code for annotations. These consist of anno elements and annotations (possibly I'll later rename these to annotations and annotation_link respectively).
This involves two new files (tg_anno.c and tg_anno.h), plus support
for the anno_t and anno_ele_t data structure in tg_iface* and
tg_cache.c.

So far anno_t is largely unimplemented, but anno_ele_t is the guts of
it and where the link to sequences or contigs occur. Right now this
means we support single tags in a similar style to Gap4, without the
link concept.

Annotations are indexed in the same bin range arrays as sequences,
utilising a flag to distinguish between seq_t and anno_ele_t
references. This code needs more work as range_t should become a
union (partly done, but commented out in tg_struct.h). This is being
checked in now as to do this properly requires a file format change.

The contig_seqs_in_range function had been extended to be part of a
family, now including contig_anno_in_range and contig_items_in_range.
The latter of these can return both sequences and annotations in one
query.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1588 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1