Overhauled the contig deregistration system. Specifically in how we handle removal from the hache table.
Also changed the meaning of contig 0. This is not a way to broadcast
to all contigs (although we still have a broadcast_event function for
this purpose) but rather a registration list for windows that wish to
receive events that take place on any contig. (Such as the contig
selector or Results Manager.)


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1598 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1