*INCOMPATIBILITY* Major overhaul on encoded data formats.
Every record written to disk now starts with a byte for the data type
and a byte for the encoding format (starting at zero for now). Every
time we read a record we assert that the type and format are correct.

This should help spot potential data corruptions before they take
place. The error handling could be improved, but assert/abort is
sufficient for now and (in theory) these should never happen anyway.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1678 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1