Added experimental code to support liblzma instead of zlib for compression of data. The results vary based on input data, but appear to be around 15% smaller at a cost of 2-4x CPU usage.
Also added code to reorder quality values and names when writing a seq
block so that sequences from the same library are adjacent to one
another. This further improves compression ratio, depending on whether
mixed libraries are in use: about 7% smaller on a 1000genomes mixed
sample.

Both of these changes are hidden behind #ifdefs and are currently disabled.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1927 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1