Removed the LD_LIBRARY_PATH definitions in global.mk (they wouldn't work on MacOS X anyway) and added -Wl,-rpath-link,$(L) to the linker flags instead.
Similarly added -rpath-link for io_lib's install directory. This is
because typically it may be built along side the staden package with
neither yet installed or without the user yet setting LD_LIBRARY_PATH
up.

This also has provision to switching to -rpath later if desired,
although rpath is - of course - evil.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1972 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1