Resurrected (and majorly rewrote, hence no real history) the old staden.profile script.
It's not required to source this, but some programs tightly linking
with the staden package or people wishing to run the smaller tools
rather than the main GUI apps will need to get their environment set
up properly: LD_LIBRARY_PATH, MANPATH, etc.

This script does that and so should be sourced from a bourne-shell
syntax login script as desired.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@1995 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1