Control-a and control-e now look at the cutoffs editor parameter to determine whether they should jump to the visible start/end of the sequence/contig or the hidden start/end.
In a similar vein, the left/right arrow keys no longer scroll past the
visible end either unless cutoffs are shown even when complemented
(previously this was bugged).

To simplify the implementation of this, I added a new function
sequence_get_orient() which returns the relative orientation of this
sequence vs the contig. This is more than just "s->len < 0" now
considering the complemented status of bins en-route to the root node
and so fixes a bug the bug in scrolling along sequences in a
complemented contig.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2019 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1