1) More work on undo, especially when undoing a base deletion that is the first/last base in hidden data, at the junction of clipped vs used. Normally when we manually insert here we want the base to be in the used visible portion, but if we're inserting as a result of undoing a deletion then it's probably cutoff data we're adding back instead.
This involved changing sequence_get_base to return whether a base is
in used or cutoff data, along with everything that calls it.

Lots of cursor work involving cutoff data. When we turn off cutoffs
and our cursor is in the cutoff region of a sequence the cursor now
gets moved to the consensus. Moving around with arrow keys will also
skip cutoff data when cutoffs are not shown, similarly for making
selections. Finally when undoing changes that occured in the cutoff
data we now automatically make the cutoffs visible again if they were
not already.

This also involved modifying the parameters to edSetCursorPos.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2027 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1