Fixed missing contig notifications on events like complementing a contig, joining contigs or simply just saving from the editor. Now we handle these the clearing of the FIJ plot has been removed as it updates correctly.
Plots that used to register with all contigs in a verbatim manner now
register with contig 0. (And indeed in many case they were incorrect
anyway as they registered with contig 1..Ncontigs rather than the actual
contig record numbers.)

To facilitate this the send_event function has been improved so that
sending to contig -N will send the message to all plots registered
against contig zero, but claiming to been an event sent to contig N
rather than 0 so that the receiver knows the request. This change is
internal to how tg_register works and not something that is visible
outside of that file.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2030 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1