Added a CSIR_SORT_BY_SEQ_TECH flag for the contig_items_in_range function. This required adding seq_tech to the rangec_t struct (in-memory, not on-disk version) too. Should this be added to on-disk struct at some stage for efficiency?
For now the editor always passes this flag in so capillary sequences
are always on top. I need to make this adjustable.

The editor also no longer calls contig_items_in_range continually
(even on mouse overs) as it now correctly tracks edits and so caches
this data.

Finally the seq_tech field can now be (and is) displayed in the brief
status line of the editor.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2157 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1