Updated the way we interpret consensus tags.
Previously the range (flags & GRANGE_FLAG_TAG_SEQ) bit was used to
distinguish sequence tags vs consensus tags, but also the range
pair_rec held the record number the tag was bound to. This makes sense
for sequences, but we also stored the consensus record for consensus
tags.

The problem comes when breaking a contig. We reparent an entire
section of bin tree, attaching it to a new contig record. We do not
wish to go through the entire bin tree changing r->pair_rec to be the
new contig ID. Fortunately checking the GRANGE_FLAG_TAG_SEQ bit is
sufficient to know if the tag is for the consensus instead.

So... changed cases where we checked r->pair_rec == contig_rec to
instead use !(r->flags & GRANGE_FLAG_TAG_SEQ).


Also updated the GFF export code in export_contigs.c so it does hex
escaping instead of C-string escaped (as mandated in GFF3 spec at
http://www.sequenceontology.org/gff3.shtml).


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2165 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1