Mass tidy up of cache_incr and cache_decr calls. This fixes several bugs which often revealed themselves as "cache_unload: Assertion `io->base || ci->updated == 0' failed."
This happens when we cache_decr too much, making the ref count go to
zero and removing the item from the hachetable in_use list. The in_use
list is scanned through looking for ci->updated==1 when flushing
changes, so a record which has been modified and then decrememented
doesn't get written on a flush. Instead it's identified in the above
insert upon closing the io handle.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2234 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1