Mid-level item deallocation:
Cache items now have a forgetme bit-flag. When set it indicates the
record should be removed. The cache_rec_deallocate function does this.
This is turn then calls the 'iface' destroy methods.

Added asserts to cache_decr. We should never get to negative ref
counts. Equally so we should never hit a zero ref count on a modified
item. (This was the cause of a corruption recently.)


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2240 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1