Fixed, I hope, a reference counting problem.
When running cache_rw on a child io cache for an item which is blocked
(ie seq or anno), the first time through we create a new (eg) SeqBlock
record. Subsuquent cache_rw on other sequences within the same block
were not updating the reference count on the cache_master object, but
cache_decr did still want to decrememnt them.

This shows up as an assertion failure if you try to edit two
neighbouring sequences (I tried adjusting clip points, but I expect it
happens with other edits too).


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2254 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1