Major speed improvements to Find Internal Joins and Find Repeats.
FIJ is around 10x faster in quick mode, and I also added a fastest
mode that takes a few more liberties in order to speed things up
further (another 2-3x over quick mode). Speed gains will vary by size
of project and number of contigs. The above was on a project with
~1500 contigs and ~20Mb of consensus.

Some related speed ups may be possible in the sensitive mode, but for
now I've left that code largely as-is.

One of the speeds to FIJ involved the hashing code, which is the core
of Find Repeats. Consequentially Find Repeats is somewhere around 100x
faster! Again, depending on the project.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2462 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1