In attempts to track down crashes when inserting columns of pads to contigs that have been highly broken, complemented and joined I temporary implemented a new version of contig_insert_base(). For now this is #ifdef-ed out though as it turns out the cause is a disparity between s->len and r->end - r->start + 1: cause unknown!
So, added some asserts to check for this. This doesn't fix the bug and
data corruption! It just means it'll hopefully it'll be pointed out
far sooner after it occurs, giving a more obvious link between cause
and effect.

Also improved robustness of sequence insertion to check for attempting
to insert off the end of a sequence.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2493 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1