Fixed a bug where the tcl interface to sequence_insert_bases and sequence_delete_bases didn't update the range_t structure. Consequentially the sequence length and the start..end length in the range struct could differ.
This was detected via the assertions added in svn update #2493, so we
now know what was altering the data that triggered the bug fixed
there.

From C the only other use of the sequence ins/del functions are from
the code in tg_contig.c to insert/delete entire columns of pads, and
these already updated the range arrays correctly. Resolved the issue
by creating a new sequence_range_length() function to update the range
data any time we feel it's been invalidated due to changing sequence
length.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2510 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1