Fixed two bugs with the contig iterators, introduced in the recent change to try and speed them up.
Firstly, CITER_SMALL_BS had a value which clashed with CITER_IEND
causing data to always be sorted on right hand end even when
explicitly asked to sort of left end. This means iterators could skip
data.

Secondly, we jumped CITER_BS into the contig for each range search,
even when the first one only fetched CITER_SMALL_BS worth of data.

The affects were ride ranging, but mostly hard to detect. The most
obvious impact is control-a and control-e to jump to the start and end
of the contig in the contig editor would sometimes not take us all the
way there when faced with widely varying sequence lengths (eg
capillary assemblies).


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2540 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1