Lots of overhauls to editor Undo, which also involved adding a new move_seq method to the tg-tcl interface.
These fix undo of removal of pad columns for the first or last base in
a contig and the first or last base in a sequence. They also fix undo
of tag deletion - we now recreate the tag using the old record number
instead of a new one (which broke subsequent undo ops on that same
tag).


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2585 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1