Fixed wide spread false meme about anno_ele_t struct. The idx field is an index into the anno_ele_block array, but in many places it was used in a manner as if it was the bin_index (ie like seq_t->bin_index vs seq_t->idx).
There is no bin_index in the anno_ele_t struct, rightly or wrongly.
Rather than simply remove the code I changed it to access a->bin_idx
(which doesn't exist yet) and commented it out so I can add the code
back to update this if and when I extend the structure to contain the
bin index.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2654 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1