I now have a better undertstanding of the causes of bin consensus sequences being invalid, so removed the warning.
The cached consensus will be correct, just not always spanninig the
entire bin. This can happen when bin_add_range adds the cached cons to
a larger bin than the one we were computing it for. We could of course
force it to only ever add to the bin we expected, and I tried this.
For some reason it caused yet more errors though, so for now I'm
glossing over this by simply gracefully handling the case.

It's not as fficient as it could be therefore as it sometimes has to
recompute the consensus again as the bin has grown, but it's
sufficient and I believe reliable.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2683 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1