Major rewrite of contig_insert_base and contig_delete_base.
These now take into account the sequence cutoff data and don't insert
or delete within those portions. The implementation is considerably
trickier to get right however. Eg on insertion:

1) It is no longer the case that all tags to the right get shifted to
the right. If they're right-hand cutoff data they will not.

2) It is no longer the case that all tags to the left will not get
shifted right. If they're in left-hand cutoff data and the read starts
to the right, then the read will be moved right causing the left tags
to move too.

3) Inserting a column of pads may not actually grow the contig
(likewise deleting may not shrink it). If the contig extents were
defined by cutoff data of a sequence and we're inserting a column that
doesn't insert into the sequence that is furhest right (so we're
inserting at a pos covering the right-hand cutoff data of it) then the
contig will not grow.

These and other problems lead to the requirement that tags should no
longer be created within the cutoff regions of the consensus.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2697 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1