Improved sequence name searching as per RT#231099
When looking for a sequence name/number we now jump to the first
visible (unclipped) base of the sequence instead of the first clipped
base.  To implement this I've created a new
sequence_get_clipped_position function to replace the
sequence_get_position version.

I also changed the editor X position so that new searches that have to
reposition the scrollbar to make the match visible does so with the
match 10 bases in from the left edge instead of the right edge as before.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2706 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1