When displaying the editor we now call update_library_stats regularly on reads with libraries in order to check consistency. This in turn does cache_rw and attempts to cache_dup from base io to child io, causing a "FIXME: implement library_dup" warning message as this hadn't been implemented yet.
However it also causes another side effect - the editor thinks we've
modified something and asks if we want to save; annoying too. For now
I just changed the cache_rw on libraries to always work on the base io
even if we're trying to cache_rw on the child io, solving both issues
at once. It's not ideal from a code purist view, but it works.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2757 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1