Added Check Assembly to Gap5. It doesn't have the option of Gap4 to align cutoff data to the consensus, but that is best served in a different route now. Instead this implementation is simply comparing used/visible data against the consensus over a window length to find highly discrepant regions of sequences.
Cutoff plan: I observe lots of places where a whole raft of data
has been clipped back to make it align. For individual sequences this
probably isn't an issue - maybe it's chimeric or poor quality - but if
multiple sequences have been clipped at the same position and the
clipped data matches then it is indicative of a problem such as a
collapsed repeat.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2763 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1