Major join editor bug fix.
When making a join, the first thing the editor does is to save both
contigs. This causes a redraw of both editors and in doing so was
causing the editors to adjust the X scrollbar to ensure the editing
cursor is visible on screen. If the join Lock mode wasn't enable and
the editor cursors weren't both visible this would shift one editor
with respect to the other, immediately before making the join. It
happens so fast you can't see it.

The effect was a catastrophically broken join with sequences out of
alignment.

The fix was changing the REG_LENGTH event callback in the editor to
not request the cursor to be visible when calling $ed set_cursor.

I also added an explicit recalculation of xx->link->lockOffset in
edJoin as it assumes this to be valid, but it appears something
already recomputes this (maybe the save). I left that fix there though
for belt-n-braces extra checking.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2786 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1