Tag directions are now stored in the database, and can be edited/set using the Tag Editor within the contig editor.
This required a change of database format, with the master format
number now being 4, meaning that older gap5s will not be able to load
the latest databases (but vice versa is true).

I also added the ability of tg_index to create databases in older
formats. In this case run "tg_index -v 3". However doing so will mean
attempts to set tag strands will silently fail as they cannot be
written back to disc.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2793 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1