Changed unpadded GFF coordinates to agree with the new unpadded coords in the editor: base 1 is the left base, regardless of where the contig start coordinate is.
This works well for local users and it matches up with coordinates in
a FASTA file, but may cause issues for people using gap5 in a
reference coordinate system. (Right now this isn't common, if at all,
given there are better viewers for such things. If it becomes a
problem we can revisit this and make the unpadded coordinate start
point optional.)


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2819 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1