Removed the defunct io-reg.h, which in most places was being included but ignored due to the correct tg_register.h having replaced it already.
The exception to this is in restriction_enzymes.c which was only
including the old io-reg.h, and it turns out in doing so was making
some registration handling bugs due to int vs tg_rec on for record
IDs.

Also added RENAME as a registration type. This is sent when a contig
changes name.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2827 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1