Fixed some B+Tree update errors when dealing with large numbers of duplicate strings (more than an entire nodes worth).
Fixed a tree error when updating the parent after a deletion of the
right-most entry in a node. I'm not sure if this was capable of
recreating the issues users sometimes see (entries going missing from
the tree), but it's an inconsistency which may lead to more errors
downstream.

Also improved debugging with -DTEST_MAIN.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2913 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1