Major addition (and corresponding database format version change) involving caching of read pair data.
The Bin GRange arrays now hold pair_{start, end, mqual, contig} in
addition to the previously held pair_rec value. This means the
absolute position of a read-pair can be obtained fast (faster even
than the absolute position of "yourself") without needing to resort to
extra seeks and I/O.

However this data can become outdated if edits occur, such as
breaking/joining, complementing or even just insertion and deletion to
contigs and/or sequences. So every range element also now has a
pair_timestamp indicating when the pair_* records were last validated.

The timestamp is compared against the pair_contig, if it still exists,
and against the main database timestamp. (The timestamps are simply
incremental numbers rather than actual times.)

There is a high chance of an error in this caching code given the
complexity and number of places that we could invalidate the cache,
but it has been well tested.

Benchmarks
==========

After initial tg_index (which sets valid timestamps) I see approx 100x
reduction in seeks and I/O for the template display in "Acc"urate mode
(this is now the default mode). ~1.1k vs 107k seeks.

After editing a contig, scrolling in the template display then
drastically slows down while the data is recomputed, but not by as
much as the old template display was at all times. It's maybe 5-6x
slower than a fully up to date version, and soon speeds up again.



git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@3083 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1