Fixed iteration on wrong contig number in contig_register_join.
Iterator hit searches on cto, but the next search incorrectly used cfrom.
This could lead to hit returning entries for cfrom if cto and cfrom happened
to hash into the same bin.  If it did, they were incorrectly deleted which
could lead to a crash later on in contig_reg_remove when it tries to
access the deleted item.  Changing HacheTableNext to use cto fixes this.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@3272 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1